Move all analytics files to /src/analytics
