+++ config/webpack.config.js
@@ -55,10 +55,10 @@ const commonConfig = ({ dev }) => {
       alias: {
         ...searchIgnoredStyles(path.resolve(__dirname, '../')),
         ...imageNullLoader(),
-        // do not consume unfetch from nested dependencies
-        unfetch: path.resolve(__dirname, '../src/js/unfetch'),
         // charts override for the PDF renderer
         '@patternfly/react-charts/dist/js/components/ChartUtils/chart-theme': path.resolve(__dirname, '../src/js/overrides/chart-utils-override.js'),
+        // do not consume unfetch from nested dependencies
+        unfetch: path.resolve(__dirname, '../src/moduleOverrides/unfetch'),
         '@scalprum/core': path.resolve(__dirname, '../node_modules/@scalprum/core'),
         '@scalprum/react-core': path.resolve(__dirname, '../node_modules/@scalprum/react-core'),
       },
