+++ src/analytics/SegmentContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+import { AnalyticsBrowser } from '@segment/analytics-next';
+
+const SegmentContext = createContext<{ ready: boolean; analytics?: AnalyticsBrowser }>({
+  ready: false,
+  analytics: undefined,
+});
+
+export default SegmentContext;
