+++ src/analytics/useSegment.ts
@@ -1,0 +1,14 @@
+import { useContext } from 'react';
+import { useIntl } from 'react-intl';
+
+import messages from '../js/Messages';
+import SegmentContext from './SegmentContext';
+
+export function useSegment() {
+  const intl = useIntl();
+  const ctx = useContext(SegmentContext);
+  if (!ctx) {
+    throw new Error(`${intl.formatMessage(messages.segmentError)}`);
+  }
+  return ctx;
+}
